% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aCML}
\alias{aCML}
\title{Atypical chronic myeloid leukemia dataset}
\format{TRONCO compliant dataset}
\source{
data from http://www.nature.com/ng/journal/v45/n1/full/ng.2495.html
}
\usage{
data(aCML)
}
\description{
This file contains a TRONCO compliant dataset
}
\author{
Luca De Sano
}

