% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{AND}
\alias{AND}
\title{AND}
\usage{
AND(...)
}
\arguments{
\item{...}{Atoms of the co-occurance pattern given either as labels or as partielly lifted vectors.}
}
\value{
Vector to be added to the lifted genotype resolving the co-occurance pattern
}
\description{
AND hypothesis
}
