% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{annotate.description}
\alias{annotate.description}
\title{annotate.description}
\usage{
annotate.description(x, label)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{label}{A string}
}
\value{
A TRONCO compliant dataset.
}
\description{
Annotate a description on the selected dataset
}
\examples{
data(test_dataset)
annotate.description(test_dataset, 'new description')

}

