% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{annotate.stages}
\alias{annotate.stages}
\title{annotate.stages}
\usage{
annotate.stages(x, stages, match.TCGA.patients = FALSE)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{stages}{A list of stages. Rownames must match samples list of x}

\item{match.TCGA.patients}{Match using TCGA notations (only first 12 characters)}
}
\value{
A TRONCO compliant dataset.
}
\description{
Annotate stage information on the selected dataset
}
\examples{
data(test_dataset)
data(stage)
test_dataset = annotate.stages(test_dataset, stage)
as.stages(test_dataset)
}

