% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{as.bnlearn.network}
\alias{as.bnlearn.network}
\title{as.bnlearn.network}
\usage{
as.bnlearn.network(x, model = names(as.models(x))[1], makeValid = TRUE)
}
\arguments{
\item{x}{A reconstructed model (the output of tronco.capri or tronco.caprese)}

\item{model}{The name of the selected regularization}

\item{makeValid}{Transform the bootstrapped data into a valid 2-categories input data}
}
\description{
Convert a TRONCO object in a Bnlearn network.
}
\examples{
data(test_model)
as.bnlearn.network(test_model)

}

