% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.colors}
\alias{as.colors}
\title{as.colors}
\usage{
as.colors(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A named vector of colors.
}
\description{
Return the colors associated to each type of event in 'x', which should be a
TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
as.colors(test_dataset)

}

