% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.dataset}
\alias{as.dataset}
\title{as.dataset}
\usage{
as.dataset(x)
}
\arguments{
\item{x}{A TRONCO model.}
}
\value{
A TRONCO model without models, hypothesis and patterns.
}
\description{
Remove models, hypothesis and patterns from a TRONCO model.
}

