% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.error.rates}
\alias{as.error.rates}
\title{as.error.rates}
\usage{
as.error.rates(x, models = names(x$model))
}
\arguments{
\item{x}{A TRONCO model.}

\item{models}{A subset of reconstructed models, all by default.}
}
\value{
The estimated rates of false positives an negatives in the data, given the model.
}
\description{
Extract the estimated rates of false positives an negatives in the data, given the model.
A subset of models if multiple reconstruction have been performed can be extracted.
}
\examples{
data(test_model)
as.error.rates(test_model)
}

