% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.events.in.sample}
\alias{as.events.in.sample}
\title{as.events.in.sample}
\usage{
as.events.in.sample(x, sample)
}
\arguments{
\item{x}{A TRONCO compliant dataset}

\item{sample}{Vector of sample names}
}
\value{
A list of events which are observed in the input samples list
}
\description{
Return a list of events which are observed in the input samples list
}

