% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.events}
\alias{as.events}
\title{as events}
\usage{
as.events(x, genes = NA, types = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{genes}{The genes to consider, if NA all available genes are used.}

\item{types}{The types of events to consider, if NA all available types are used.}
}
\value{
A matrix with 2 columns (event type, gene name) for the events found.
}
\description{
Return all events involving certain genes and of a certain type in 'x', which should be a
TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(maf)
mutations = import.MAF(maf)
as.events(mutations)
as.events(mutations, types='Mutation')
as.events(mutations, genes = 'ABAT')
as.events(mutations, types='Missing')
}

