% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as.events.test}
\alias{as.events.test}
\title{as events matrix}
\format{matrix}
\source{
fake data
}
\usage{
data(as.events.test)
}
\description{
This data set list ...
}
\author{
Luca De Sano
}

