% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.genes.hypotheses}
\alias{as.genes.hypotheses}
\title{as.genes.hypotheses}
\usage{
as.genes.hypotheses(x, hypotheses = NULL)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{hypotheses}{????????}
}
\value{
????????
}
\description{
Return ???????? RIDENOMINARLA CON PATTERNS E DARLE UN NOME PIU INTELLEGIBILE?
}

