% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.genes.in.patterns}
\alias{as.genes.in.patterns}
\title{as.genes.in.patterns}
\usage{
as.genes.in.patterns(x, patterns = NULL)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{patterns}{A list of patterns for which the list will be returned}
}
\value{
A list of genes present in patterns which consitute CAPRI's hypotheses
}
\description{
Return the list of genes used in patterns
}

