% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.genes}
\alias{as.genes}
\title{as.genes}
\usage{
as.genes(x, types = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{types}{The types of events to consider, if NA all available types are used.}
}
\value{
A vector of gene symbols for which a certain type of event exists
}
\description{
Return all gene symbols for which a certain type of event exists in 'x', which should be a
TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
as.genes(test_dataset)

}

