% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.genotypes}
\alias{as.genotypes}
\title{as genotypes}
\usage{
as.genotypes(x)
}
\arguments{
\item{x:}{the dataset.}

\item{genes:}{a list of genes to query, if NA all genes are used.}
}
\description{
as genotypes
}

