% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.hypotheses}
\alias{as.hypotheses}
\title{Return the name of hypotheses in the dataset}
\usage{
as.hypotheses(x)
}
\arguments{
\item{x}{the dataset.}
}
\description{
Return the name of hypotheses in the dataset
}

