% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.joint.probs}
\alias{as.joint.probs}
\title{as.joint.probs}
\usage{
as.joint.probs(x, events = as.events(x), models = names(x$model),
  type = "observed")
}
\arguments{
\item{x}{A TRONCO model.}

\item{events}{A subset of events as of \code{as.events(x)}, all by default.}

\item{models}{A subset of reconstructed models, all by default.}

\item{type}{Either observed ('observed') or fit ('fit') probabilities, 'observed' by default.}
}
\value{
The joint probabilities in a TRONCO model.
}
\description{
Extract the joint probabilities from a TRONCO model. The return matrix is indexed with rownames/colnames which
represent genotype keys - these can be resolved with function \code{keysToNames}. It is possible to
specify a subset of events to build the matrix, a subset of models if multiple reconstruction have
been performed. Also, either the observed or fit probabilities can be extracted.
}

