% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.kfold.eloss}
\alias{as.kfold.eloss}
\title{as.kfold.eloss}
\usage{
as.kfold.eloss(x, models = names(x$model))
}
\arguments{
\item{x}{A TRONCO model.}

\item{models}{A subset of reconstructed models, all by default.}
}
\value{
All the bootstrap scores in a TRONCO model
}
\description{
Returns a dataframe with all the entropy loss score in a 
TRONCO model. It is possible to specify models if multiple
reconstruction have been performed.
}
\examples{
data(test_model)
as.kfold.eloss(test_model)
as.kfold.eloss(test_model, models='aic')

}

