% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.models}
\alias{as.models}
\title{as.models}
\usage{
as.models(x, models = names(x$model))
}
\arguments{
\item{x}{A TRONCO model.}

\item{models}{The name of the models to extract, e.g. 'bic', 'aic', 'caprese', all by default.}
}
\value{
The models in a reconstructed object.
}
\description{
Extract the models from a reconstructed object.
}
\examples{
data(test_model)
as.models(test_model)

}

