% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.name}
\alias{as.name}
\title{as.name}
\usage{
as.name(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The name annotating the dataset, if any.
}
\description{
Return the name annotating the dataset, if any. Input 'x' should be
a TRONCO compliant dataset - see \code{is.compliant}.
}

