% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.parents.pos}
\alias{as.parents.pos}
\title{as.parents.pos}
\usage{
as.parents.pos(x, events = as.events(x), models = names(x$model))
}
\arguments{
\item{x}{A TRONCO model.}

\item{events}{A subset of events as of \code{as.events(x)}, all by default.}

\item{models}{A subset of reconstructed models, all by default.}
}
\value{
A list of parents for each node
}
\description{
Get parents for each node
}

