% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.patterns}
\alias{as.patterns}
\title{as.patterns}
\usage{
as.patterns(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The patterns in the dataset which constitute CAPRI's hypotheses.
}
\description{
Return the patterns in the dataset which constitute CAPRI's hypotheses.
}
\examples{
data(test_dataset)
as.patterns(test_dataset)

}

