% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.stages}
\alias{as.stages}
\title{as.stages}
\usage{
as.stages(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A matrix with 1 column annotating stages and rownames as sample IDs.
}
\description{
Return the association sample -> stage, if any. Input 'x' should be a
TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
data(stage)
test_dataset = annotate.stages(test_dataset, stage)
as.stages(test_dataset)

}

