% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as.stages.test}
\alias{as.stages.test}
\title{as stages matrix}
\format{matrix}
\source{
fake data
}
\usage{
data(as.stages.test)
}
\description{
This data set list ...
}
\author{
Luca De Sano
}

