% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.summary.statistics}
\alias{as.summary.statistics}
\title{as.summary.statistics}
\usage{
as.summary.statistics(x, events = as.events(x), models = names(x$model))
}
\arguments{
\item{x}{A TRONCO model.}

\item{events}{A subset of events as of \code{as.events(x)}, all by default.}

\item{models}{A subset of reconstructed models, all by default.}
}
\value{
All the statistics for a TRONCO model
}
\description{
Returns a dataframe with all the statistics for a 
TRONCO model. It is possible to specify a subset of events or 
models if multiple reconstruction have been performed.
}
\examples{
data(test_model)
as.summary.statistics(test_model)
as.summary.statistics(test_model, events=as.events(test_model)[5:15,])

}

