% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.types}
\alias{as.types}
\title{as types}
\usage{
as.types(x, genes = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{genes}{A list of genes to consider, if NA all genes are used.}
}
\value{
A matrix with 1 column annotating stages and rownames as sample IDs.
}
\description{
Return the types of events for a set of genes which are in 'x', which should be a
TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(maf)
mutations = import.MAF(maf, sep=';')
as.types(mutations)
as.types(mutations, genes = 'ABAT')
}

