% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.bootstrap.R
\name{bootstrap.capri}
\alias{bootstrap.capri}
\title{bootstrap.capri}
\usage{
bootstrap.capri(dataset, hypotheses, command.capri, regularization, do.boot,
  nboot.capri, pvalue, min.boot, min.stat, boot.seed, do.estimation, silent,
  reconstruction, command = "non-parametric", nboot = 100,
  bootstrap.statistics = list(), verbose = FALSE)
}
\arguments{
\item{dataset}{a dataset describing a progressive phenomenon}

\item{hypotheses}{a set of hypotheses referring to the dataset}

\item{command.capri}{type of search, either hill climbing (hc) or tabu (tabu)}

\item{regularization}{regularizators to be used for the likelihood fit}

\item{do.boot}{should I perform bootstrap? Yes if TRUE, no otherwise}

\item{nboot.capri}{integer number (greater than 0) of bootstrap sampling to be performed}

\item{pvalue}{pvalue for the tests (value between 0 and 1)}

\item{min.boot}{minimum number of bootstrapping to be performed}

\item{min.stat}{should I keep bootstrapping untill I have nboot valid values?}

\item{boot.seed}{seed to be used for the sampling}

\item{do.estimation}{should I perform the estimation of the error rates and probabilities?}

\item{silent}{should I be verbose?}

\item{reconstruction}{todo}

\item{bootstrap.statistics}{todo}

\item{verbose}{todo}

\item{command:}{should I perform non-parametric or parametric bootstrap?}

\item{nboot:}{number of bootstrap resampling to be performed}
}
\value{
bootstrap.statistics: statistics of the bootstrap
}
\description{
perform non-parametric or parametric bootstrap to evalutate the confidence of the reconstruction
}

