% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_get_dag_scores}
\alias{C_get_dag_scores}
\title{execute the get.dag.scores procedures}
\usage{
C_get_dag_scores(dataset, adj_matrix, epos, eneg)
}
\arguments{
\item{dataset}{A TRONCO dataset}

\item{adj_matrix}{An adj matrix}

\item{epos}{false positive rates}

\item{eneg}{false negative rates}
}
\description{
execute the get.dag.scores procedures
}

