% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.algorithm.R
\name{caprese.fit}
\alias{caprese.fit}
\title{caprese.fit}
\usage{
caprese.fit(dataset, lambda = 0.5, do.estimation = FALSE, silent = FALSE)
}
\arguments{
\item{dataset}{a dataset describing a progressive phenomenon}

\item{lambda}{shrinkage parameter (value in [0,1])}

\item{do.estimation}{should I perform the estimation of the error rates and probabilities?}

\item{silent}{execute the algorithm in silent mode}
}
\value{
topology: the reconstructed tree-like topology
}
\description{
reconstruct the best tree-like topology
}

