% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{capri.fit}
\alias{capri.fit}
\title{capri.fit}
\usage{
capri.fit(dataset, hypotheses = NA, command = "hc",
  regularization = c("bic", "aic"), do.boot = TRUE, nboot = 100,
  pvalue = 0.05, min.boot = 3, min.stat = TRUE, boot.seed = NULL,
  do.estimation = FALSE, silent = FALSE)
}
\arguments{
\item{dataset}{a dataset describing a progressive phenomenon}

\item{hypotheses}{hypotheses to be considered in the reconstruction}

\item{command}{type of search for the likelihood fit, either hill climbing (hc) or tabu (tabu)}

\item{regularization}{regularizators to be used for the likelihood fit}

\item{do.boot}{should I perform bootstrap? Yes if TRUE, no otherwise}

\item{nboot}{integer number (greater than 0) of bootstrap sampling to be performed}

\item{pvalue}{pvalue for the tests (value between 0 and 1)}

\item{min.boot}{minimum number of bootstrapping to be performed}

\item{min.stat}{should I keep bootstrapping untill I have nboot valid values?}

\item{boot.seed}{seed to be used for the sampling}

\item{do.estimation}{should I perform the estimation of the error rates and probabilities?}

\item{silent}{should I be verbose?}
}
\value{
topology: the reconstructed tree topology
}
\description{
reconstruct the best dag topology running CAPRI algorithm
}

