% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{cbio.query}
\alias{cbio.query}
\title{cbio.query}
\usage{
cbio.query(cbio.study = NA, cbio.dataset = NA, cbio.profile = NA,
  genes, file = NA)
}
\arguments{
\item{cbio.study}{Cbio study ID}

\item{cbio.dataset}{Cbio dataset ID}

\item{cbio.profile}{Cbio genetic profile ID}

\item{genes}{A list of < 900 genes to query}

\item{file}{String containing filename for RData output. If NA no output will be provided}
}
\value{
A list with two dataframe: the gentic profile required and clinical data for the Cbio study.
}
\description{
Wrapper for the CGDS package to query the Cbio portal. This can work either automatically, if one
sets \code{cbio.study}, \code{cbio.dataset} or \code{cbio.profile}, or interactively otherwise. A
list of genes to query with less than 900 entries should be provided. This function returns a list
with two dataframe: the gentic profile required and clinical data for the Cbio study. Output is also
saved to disk as Rdata file. See also http://www.cbioportal.org.
}
