% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{change.color}
\alias{change.color}
\title{change.color}
\usage{
change.color(x, type, new.color)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{type}{An event type}

\item{new.color}{The new color (either HEX or R Color)}
}
\value{
A TRONCO complian dataset.
}
\description{
Change the color of an event type
}
\examples{
data(test_dataset)
dataset = change.color(test_dataset, 'ins_del', 'red')

}

