% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{check.dataset}
\alias{check.dataset}
\title{check.dataset}
\usage{
check.dataset(dataset, adj.matrix, verbose)
}
\arguments{
\item{dataset}{a dataset describing a progressive phenomenon}

\item{adj.matrix}{adjacency matrix of the topology}

\item{verbose}{should I print the warnings? Yes if TRUE, no otherwise}
}
\value{
valid.dataset: a dataset valid accordingly to the probability raising
}
\description{
check if the dataset is valid accordingly to the probability raising
}

