% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confidence.R
\name{confidence}
\alias{confidence}
\alias{confidence.conditional}
\alias{confidence.data.conditional}
\alias{confidence.data.joint}
\alias{confidence.data.single}
\alias{confidence.fit.conditional}
\alias{confidence.fit.joint}
\alias{confidence.fit.single}
\alias{confidence.joint}
\alias{confidence.single}
\title{provides various kinds of confidence measures for an inferred progression model}
\usage{
confidence.data.joint(topology)

confidence.fit.joint(topology)

confidence.data.single(topology)

confidence.fit.single(topology)

confidence.data.conditional(topology)

confidence.fit.conditional(topology)

confidence.single(topology)

confidence.joint(topology)

confidence.conditional(topology)
}
\arguments{
\item{topology}{A topology returned by the reconstruction algorithm}
}
\description{
A set of functions to visualise and compare the probability of each event in the progression model, as well as their joint and conditional distributions. These can be evaluated both in the data (observed probabilities) and in the reconstructed model (fitted probabilities).
}
\details{
\code{confidence.data.joint} plot the pairwise observed joint probability of the events

\code{confidence.fit.joint} plot the pairwise fitted joint probability of the events

\code{confidence.data.single} plot the observed probability of each event

\code{confidence.fit.single} plot the fitted probability of each event

\code{confidence.data.conditional} plot the pairwise observed conditional probability of the events

\code{confidence.fit.conditional} plot the pairwise fitted conditional probability of the events

\code{confidence.single} plot the difference between the observed and fitted  probability of each event

\code{confidence.joint} plot the pairwise difference between the observed and fitted joint probability of the events

\code{confidence.conditional} plot the pairwise difference between the observed and fitted conditional probability of the events
}

