% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{consolidate.data}
\alias{consolidate.data}
\title{consolidate.data}
\usage{
consolidate.data(x, print = FALSE)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{print}{A boolean value stating whether to print of not the summary}
}
\value{
The list of any 0 probability, 1 probability and indistinguishable.
}
\description{
Verify if the input data are consolidate, i.e., if there are events
with 0 or 1 probability or indistinguishable in terms of
observations
}
\examples{
data(test_dataset)
consolidate.data(test_dataset)

}
