% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crc_gistic}
\alias{crc_gistic}
\title{GISTIC example data}
\format{GISTIC score}
\source{
data from http://www.nature.com/nature/journal/v487/n7407/full/nature11252.html
}
\usage{
data(crc_gistic)
}
\description{
This dataset contains an example of GISTIC input of a crc cohort of patients
}
\author{
Daniele Ramazzotti
}

