% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crc_maf}
\alias{crc_maf}
\title{MAF example data}
\format{Manual Annotated Format}
\source{
data from http://www.nature.com/nature/journal/v487/n7407/full/nature11252.html
}
\usage{
data(crc_maf)
}
\description{
This dataset contains an example of MAF input of a crc cohort of patients
}
\author{
Daniele Ramazzotti
}

