% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crc_plain}
\alias{crc_plain}
\title{Plain mutation dataset}
\format{plain data}
\source{
data from http://www.nature.com/nature/journal/v487/n7407/full/nature11252.html
}
\usage{
data(crc_plain)
}
\value{
A plain input
}
\description{
This dataset contains an example of plain input of a crc cohort of patients
}
\author{
Daniele Ramazzotti
}
