% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.load}
\alias{data.load}
\title{load a dataset (binary matrix) from a file or a preloaded dataset.}
\usage{
data.load(data.input)
}
\arguments{
\item{data.input}{The input file path. or a dataset loaded by \code{data} function}
}
\description{
\code{data.load} sets a global data frame 'data.values' that contains the dataset loaded from an input file.
}
\details{
\code{data.load} loads a dataset from disk and associates all columns in the dataset to a specified event. Thus, types and events must be specified before calling this function to ensure a consistency check is performed on the input dataset (see \code{types.load}, \code{types.add},  \code{events.load}, \code{events.add} to load/add types/events).
}

