% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.bootstrap.R
\name{decimal.to.binary.dag}
\alias{decimal.to.binary.dag}
\title{decimal.to.binary.dag}
\usage{
decimal.to.binary.dag(num.decimal, num.bits)
}
\arguments{
\item{num.decimal}{decimal integer to be converted}

\item{num.bits}{number of bits to be used}
}
\value{
num.binary: binary conversion of num.decimal
}
\description{
convert an integer decimal number to binary
}

