% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.event}
\alias{delete.event}
\title{delete.event}
\usage{
delete.event(x, gene, type)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{gene}{The name of the gene to delete.}

\item{type}{The name of the type to delete.}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete an event from the dataset
}
\examples{
data(test_dataset)
test_dataset = delete.event(test_dataset, 'TET2', 'ins_del')

}
