% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.model}
\alias{delete.model}
\title{delete.model}
\usage{
delete.model(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete a reconstructed model from the dataset
}
\examples{
data(test_model)
model = delete.model(test_model)
has.model(model)

}

