% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.pattern}
\alias{delete.pattern}
\title{delete.pattern}
\usage{
delete.pattern(x, pattern)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{pattern}{A pattern name}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete a pattern and every associated hypotheses from the dataset
}
\examples{
data(test_dataset)
delete.pattern(test_dataset, pattern='XOR_TET2')
}

