% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.samples}
\alias{delete.samples}
\title{delete.samples}
\usage{
delete.samples(x, samples)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{sampels}{An array of samples name}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete samples from selected dataset
}
\examples{
data(test_dataset)
dataset = delete.samples(test_dataset, c('patient 1', 'patient 4'))
}

