% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.type}
\alias{delete.type}
\title{delete.type}
\usage{
delete.type(x, type)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{type}{The name of the type to delete.}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete an event type
}
\examples{
data(test_dataset)
test_dataset = delete.type(test_dataset, 'Pattern')

}

