% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{duplicates}
\alias{duplicates}
\title{duplicates}
\usage{
duplicates(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A subset of \code{as.events(x)} with duplicated events.
}
\description{
Return the events duplicated in \code{x}, if any. Input 'x' should be
a TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
duplicates(test_dataset)

}

