% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{ebind}
\alias{ebind}
\title{ebind}
\usage{
ebind(..., silent = FALSE)
}
\arguments{
\item{...}{the input datasets}

\item{silent}{A parameter to disable/enable verbose messages.}
}
\value{
A TRONCO complian dataset.
}
\description{
Binds events from one or more datasets, which must be defined over the same set of samples.
}

