% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{enforce.numeric}
\alias{enforce.numeric}
\title{enforce.numeric}
\usage{
enforce.numeric(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
Convert the internal reprensentation of genotypes to numeric, if not.
}
\description{
Convert the internal reprensentation of genotypes to numeric, if not.
}
\examples{
data(test_dataset)
test_dataset = enforce.numeric(test_dataset)
}

