% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{enforce.string}
\alias{enforce.string}
\title{enforce.string}
\usage{
enforce.string(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
Convert the internal reprensentation of genotypes to character, if not.
}
\description{
Convert the internal representation of genotypes to character, if not.
}

