% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.estimation.R
\name{enumerate.all.paths}
\alias{enumerate.all.paths}
\title{enumerate.all.paths}
\usage{
enumerate.all.paths(ancestor.node, child.node, parents.pos)
}
\arguments{
\item{ancestor.node}{first node of the path}

\item{child.node}{last node of the path}

\item{parents.pos}{topological connections}
}
\value{
all.paths: vector of all the paths
}
\description{
enumerate all the paths between two nodes of a DAG
}

