% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.estimation.R
\name{estimate.dag.joint.probs}
\alias{estimate.dag.joint.probs}
\title{estimate.dag.joint.probs}
\usage{
estimate.dag.joint.probs(first.node, second.node, parents.pos, marginal.probs,
  conditional.probs)
}
\arguments{
\item{first.node}{first node}

\item{second.node}{second node}

\item{parents.pos}{which event is the parent? -1 if none, a list otherwise}

\item{marginal.probs}{marginal probabilities}

\item{conditional.probs}{conditional probabilities}
}
\value{
estimated.dag.joint.probs: estimated theoretical joint probability
}
\description{
estimate the theoretical joint probability of two given nodes given the reconstructed topology
}

