% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.estimation.R
\name{estimate.dag.probs}
\alias{estimate.dag.probs}
\title{estimate.dag.probs}
\usage{
estimate.dag.probs(dataset, marginal.probs, joint.probs, parents.pos,
  error.rates)
}
\arguments{
\item{dataset}{a valid dataset}

\item{marginal.probs}{observed marginal probabilities}

\item{joint.probs}{observed joint probabilities}

\item{parents.pos}{position of the parents in the list of nodes}

\item{error.rates}{rates for the false positive and the false negative errors}
}
\value{
estimated.probs: estimated marginal, joint and conditional probabilities
}
\description{
estimate the marginal, joint and conditional probabilities given the reconstructed topology and the error rates
}

