% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.estimation.R
\name{estimate.dag.samples}
\alias{estimate.dag.samples}
\title{estimate.dag.samples}
\usage{
estimate.dag.samples(dataset, reconstructed.topology,
  estimated.marginal.probabilities, estimated.conditional.probabilities,
  parents.pos, error.rates)
}
\arguments{
\item{dataset}{a valid dataset}

\item{reconstructed.topology}{the reconstructed topology}

\item{estimated.marginal.probabilities}{estimated marginal probabilities of the events}

\item{estimated.conditional.probabilities}{estimated conditional probabilities of the events}

\item{parents.pos}{position of the parents of each node}

\item{error.rates}{error rates for false positives and false negatives}
}
\value{
probabilities: probability of each sample
}
\description{
estimate the probability of observing each sample in the dataset given the reconstructed topology
}

