% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.estimation.R
\name{estimate.tree.error.rates}
\alias{estimate.tree.error.rates}
\title{estimate.tree.error.rates}
\usage{
estimate.tree.error.rates(marginal.probs, joint.probs, parents.pos)
}
\arguments{
\item{marginal.probs}{marginal probabilities}

\item{joint.probs}{joint probabilities}

\item{parents.pos}{which event is the parent? 0 if none, a number otherwise}
}
\value{
estimated.error.rates: estimated probabilities, false positive and false negative error rates
}
\description{
estimate the error rates by "L-BFGS-B" optimization in terms of L2-error
}

