% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/events.add.R
\name{events.add}
\alias{events.add}
\title{add a new event (e.g., a missense point mutation for EGFR)}
\usage{
events.add(event.name, type.name, column.number = NA)
}
\arguments{
\item{event.name}{The event label(e.g., 'EGFR') . All event labels are strings.}

\item{type.name}{The type name of this event (e.g., 'missense point'). Type names must refer to types loaded before adding an event, a consistency check raises an error if the type name is unknown.}

\item{column.number}{The dataset column to which this event is associated. Column number must be an
integer positive value.}
}
\description{
\code{events.add} sets a global data frame 'events' that contains all the events defined. Events can be added and refined incrementally, in any order.
}
\details{
\code{events.add} allows to define one event at a time. If the event was previously defined, its definition is updated to keep track of its last definition. A consistency check is performed to ensure that the type of defined event is valid. Thus, types must be defined before events are loaded (see \code{types.add}, \code{types.load}).
}
\examples{
types.add("gain", "red")
events.add("8q+", "gain", 1)
}

