% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/events.load.R
\name{events.load}
\alias{events.load}
\title{load a set of events from file}
\usage{
events.load(data.input)
}
\arguments{
\item{data.input}{The input file path or a dataset to be validated.}
}
\description{
\code{events.load} sets a global data frame 'events' that contains all event definitions found in a specified file or dataset to be validated. This is a way to automatise calls to function \code{events.add} for a bunch of events.
}
\details{
\code{events.load} load a set of events from a given file. The input file must be structured as a CSV file, where each event is defined on a separate line in the format: eventName, typeName, columnNumber.
}
\seealso{
\code{\link{events.add}}
}

