% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{events}
\alias{events}
\title{Events collection for Ovarian cancer CGH data}
\format{An example with 7 events}
\description{
This example contains a collection of events associeted to the
Ovarian cancer CGH dataset
}

