% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection.R
\name{events.selection}
\alias{events.selection}
\title{events.selection}
\usage{
events.selection(x, filter.freq = NA, filter.in.names = NA,
  filter.out.names = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{filter.freq}{[0,1] value which constriants the minimum frequence of selected events}

\item{filter.in.names}{gene symbols which will be included}

\item{filter.out.names}{gene symbols which will NOT be included}
}
\value{
A TRONCO compliant dataset.
}
\description{
select a subset of the input genotypes 'x'. Selection can be done
by frequency and gene symbols.
}
\examples{
data(test_dataset)
dataset = events.selection(test_dataset, 0.3)
}

