% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{export.graphml}
\alias{export.graphml}
\title{export.graphml}
\usage{
export.graphml(x, file, ...)
}
\arguments{
\item{x}{A TRONCO compliant dataset}

\item{file}{Where to save the output}

\item{...}{parameters for tronco.plot}
}
\description{
Create a graphML object which can be imported in cytoscape
This function is based on the tronco.plot fuction
}
\examples{
data(test_model)
export.graphml(test_model, file='text.xml', scale.nodes=0.3)

}

