% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/external.R
\name{export.mutex}
\alias{export.mutex}
\title{export mutex}
\usage{
export.mutex(x, filename = "tronco_to_mutex", filepath = "./",
  label.mutation = "SNV", label.amplification = list("High-level Gain"),
  label.deletion = list("Homozygous Loss"))
}
\arguments{
\item{x}{The TRONCO standard object}

\item{filename}{TODO}

\item{filepath}{TODO}

\item{label.mutation}{TODO}

\item{label.amplification}{TODO}

\item{label.deletion}{TODO}
}
\description{
export mutex
}

