% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/external.R
\name{export.nbs.input}
\alias{export.nbs.input}
\title{export nbs input}
\usage{
export.nbs.input(x, map_hugo_entrez, file = "tronco_to_nbs.mat")
}
\arguments{
\item{x}{The TRONCO standard object}

\item{map_hugo_entrez}{Hugo_Symbol  - Entrez_Gene_Id}

\item{file}{TODO}
}
\description{
export nbs input
}

