% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{extract.MAF.HuGO.Entrez.map}
\alias{extract.MAF.HuGO.Entrez.map}
\title{extract.MAF.HuGO.Entrez.map}
\usage{
extract.MAF.HuGO.Entrez.map(file, sep = "\\t")
}
\arguments{
\item{file}{MAF filename}

\item{sep}{MAF separator, default \'\\t\'}
}
\value{
A mapHugo_Symbol -> Entrez_Gene_Id.
}
\description{
Extract a map Hugo_Symbol -> Entrez_Gene_Id from a MAF input file. If some genes map to ID 0
a warning is raised.
}
