% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{genes.table.plot}
\alias{genes.table.plot}
\title{genes.table.plot}
\usage{
genes.table.plot(x, name, dir = getwd())
}
\arguments{
\item{x}{A TRONCO compliant dataset}

\item{name}{filename}

\item{dir}{where to save the file}
}
\description{
Generates stacked histogram
}

