% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.bootstapped.scores}
\alias{get.bootstapped.scores}
\title{get.bootstapped.scores}
\usage{
get.bootstapped.scores(dataset, nboot, adj.matrix, min.boot = 3,
  min.stat = TRUE, boot.seed = NULL, silent)
}
\arguments{
\item{dataset}{a valid dataset}

\item{nboot}{number of bootstrap resampling to be performed}

\item{adj.matrix}{adjacency matrix of the initially valid edges}

\item{min.boot}{minimum number of bootstrapping to be performed}

\item{min.stat}{should I keep bootstrapping untill I have nboot valid values?}

\item{boot.seed}{seed to be used for the sampling}
}
\value{
scores: list structure with the scores and the data generated by bootstrap
}
\description{
compute a robust estimation of the scores using rejection sampling bootstrap
}

