% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.dag.scores}
\alias{get.dag.scores}
\title{get.dag.scores}
\usage{
get.dag.scores(dataset, adj.matrix)
}
\arguments{
\item{dataset}{a valid dataset}

\item{adj.matrix}{adjacency matrix of the initially valid edges}
}
\value{
scores: observed probabilities and prima facie scores
}
\description{
compute the observed probabilities and the prima facie scores on the dataset
}

