% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.prima.facie.causes.do.boot}
\alias{get.prima.facie.causes.do.boot}
\title{get.prima.facie.causes.do.boot}
\usage{
get.prima.facie.causes.do.boot(adj.matrix, hypotheses,
  marginal.probs.distributions, prima.facie.model.distributions,
  prima.facie.null.distributions, pvalue, dataset, marginal.probs, joint.probs,
  silent = FALSE)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the initially valid edges}

\item{hypotheses}{hypotheses to be considered}

\item{marginal.probs.distributions}{distributions of the bootstrapped marginal probabilities}

\item{prima.facie.model.distributions}{distributions of the prima facie model}

\item{prima.facie.null.distributions}{distributions of the prima facie null}

\item{pvalue}{minimum pvalue for the Mann-Whitney U tests to be significant}

\item{dataset}{a valid dataset}

\item{marginal.probs}{observed marginal probabilities}

\item{joint.probs}{observed joint probabilities}

\item{silent}{Should I be verbose?}
}
\value{
prima.facie.topology: list describing the topology of the prima facie causes
}
\description{
select the best set of prima facie causes per node
}

