% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.prima.facie.causes.no.boot}
\alias{get.prima.facie.causes.no.boot}
\title{get.prima.facie.causes.no.boot}
\usage{
get.prima.facie.causes.no.boot(adj.matrix, hypotheses, marginal.probs,
  prima.facie.model, prima.facie.null, dataset, joint.probs, silent = FALSE)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the initially valid edges}

\item{marginal.probs}{marginal probabilities}

\item{prima.facie.model}{prima facie model}

\item{prima.facie.null}{prima facie null}

\item{dataset}{a valid dataset}

\item{joint.probs}{observed joint probabilities}

\item{marginal.probs}{observed marginal probabilities}
}
\value{
prima.facie.topology: adjacency matrix of the prima facie causes
}
\description{
select the best set of prima facie causes per node without bootstrap
}

