% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.prima.facie.parents.do.boot}
\alias{get.prima.facie.parents.do.boot}
\title{get.prima.facie.parents.do.boot}
\usage{
get.prima.facie.parents.do.boot(dataset, hypotheses, nboot, pvalue, adj.matrix,
  min.boot, min.stat, boot.seed, silent)
}
\arguments{
\item{dataset}{a valid dataset}

\item{nboot}{integer number (greater than 0) of bootstrap sampling to be performed}

\item{pvalue}{pvalue for the tests (value between 0 and 1)}

\item{adj.matrix}{adjacency matrix of the initially valid edges}

\item{min.boot}{minimum number of bootstrapping to be performed}

\item{min.stat}{should I keep bootstrapping untill I have nboot valid values?}

\item{boot.seed}{seed to be used for the sampling}
}
\value{
prima.facie.parents list of the set (if any) of prima facie parents for each node
}
\description{
select the set of the prima facie parents (with bootstrap) for each node
based on Suppes' definition of causation
}

