% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.algorithm.R
\name{get.tree.parents}
\alias{get.tree.parents}
\title{get.tree.parents}
\usage{
get.tree.parents(adj.matrix, marginal.probs, joint.probs, lambda)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the valid edges}

\item{marginal.probs}{observed marginal probabilities}

\item{joint.probs}{observed joint probabilities}

\item{lambda}{shrinkage parameter (value between 0 and 1)}
}
\value{
best.parents list of the best parents
}
\description{
select at the most one parent for each node based on the probability raising criteria
}

