% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.algorithm.R
\name{get.tree.scores}
\alias{get.tree.scores}
\title{get.tree.scores}
\usage{
get.tree.scores(adj.matrix, marginal.probs, joint.probs, lambda)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the valid edges}

\item{marginal.probs}{observed marginal probabilities}

\item{joint.probs}{observed joint probabilities}

\item{lambda}{shrinkage parameter (value between 0 and 1)}
}
\value{
scores: probability raising based scores
}
\description{
compute the probability raising based scores
}

