% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gistic}
\alias{gistic}
\title{GISTIC example data}
\format{GISTIC score}
\source{
fake data
}
\usage{
data(gistic)
}
\description{
This dataset contains a standard GISTIC input for TRONCO
}
\author{
Luca De Sano
}

