% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{has.duplicates}
\alias{has.duplicates}
\title{has.duplicates}
\usage{
has.duplicates(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
TRUE if there are duplicated events in \code{x}.
}
\description{
Return true if there are duplicated events in the TRONCO dataset 'x', which should be
a TRONCO compliant dataset - see \code{is.compliant}. Events are identified by a gene 
name, e.g., a HuGO_Symbol, and a type label, e.g., c('SNP', 'KRAS')
}
\examples{
data(test_dataset)
has.duplicates(test_dataset)

}

