% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{has.stages}
\alias{has.stages}
\title{has stages}
\usage{
has.stages(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
TRUE if the TRONCO dataset has stage annotations for samples.
}
\description{
Return true if the TRONCO dataset 'x', which should be a TRONCO compliant dataset
- see \code{is.compliant} - has stage annotations for samples. Some sample stages
might be annotated as NA, but not all.
}

