% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypotheses.expansion}
\alias{hypotheses.expansion}
\title{hypotheses.expansion}
\usage{
hypotheses.expansion(input_matrix, map = list(), hidden_and = T,
  expand = T, skip.disconnected = TRUE)
}
\arguments{
\item{input_matrix}{A TRONCO adjacency matrix}

\item{map}{hypothesis name - hypothesis adjacency matrix map}

\item{hidden_and}{Should I visualize hidden and?}

\item{expand}{Should I expand the hypotheses?}

\item{skip.disconnected}{Hide disconnected node}
}
\description{
Internal function for hypotheses expansion
}

