% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.add.group}
\alias{hypothesis.add.group}
\title{hypothesis add group}
\usage{
hypothesis.add.group(
  x,
  FUN,
  group,
  pattern.cause = "*",
  pattern.effect = "*",
  dim.min = 2,
  dim.max = length(group),
  min.prob = 0,
  silent = FALSE
)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{FUN}{Type of pattern to be added, e.g., co-occurance, soft or hard exclusivity.}

\item{group}{Group of events to be considered.}

\item{pattern.cause}{Possibile causes for the pattern.}

\item{pattern.effect}{Possibile effects for the pattern.}

\item{dim.min}{Minimum cardinality of the subgroups to be considered.}

\item{dim.max}{Maximum cardinality of the subgroups to be considered.}

\item{min.prob}{Minimum probability associated to each valid group.}

\item{silent}{A parameter to disable/enable verbose messages.}
}
\value{
A TRONCO compliant object with the added hypotheses
}
\description{
Add all the hypotheses related to a group of events
}
