% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.add.homologous}
\alias{hypothesis.add.homologous}
\title{hypothesis.add.homologous}
\usage{
hypothesis.add.homologous(
  x,
  pattern.cause = "*",
  pattern.effect = "*",
  genes = as.genes(x),
  silent = FALSE
)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{pattern.cause}{Possibile causes for the pattern.}

\item{pattern.effect}{Possibile effects for the pattern.}

\item{genes}{List of genes to be considered as possible homologous. For these genes, all the types of mutations will be considered functionally equivalent.}

\item{silent}{A parameter to disable/enable verbose messages.}
}
\value{
A TRONCO compliant object with the added hypotheses
}
\description{
Add all the hypotheses related to homologou events
}
