% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.add}
\alias{hypothesis.add}
\title{hypothesis add}
\usage{
hypothesis.add(data, pattern.label, lifted.pattern, pattern.effect = "*",
  pattern.cause = "*")
}
\arguments{
\item{data}{A TRONCO compliant dataset.}

\item{pattern.label}{Label of the new hypothesis.}

\item{lifted.pattern}{Vector to be added to the lifted genotype resolving the pattern related to the new hypothesis}

\item{pattern.effect}{Possibile effects for the pattern.}

\item{pattern.cause}{Possibile causes for the pattern.}
}
\value{
A TRONCO compliant object with the added hypothesis
}
\description{
Add a new hypothesis by creating a new event and adding it to the compliant genotypes
}

