% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.connections}
\alias{hypothesis.connections}
\title{hypothesis.connections}
\usage{
hypothesis.connections(adj.matrix, hypotheses.label)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the topology}

\item{hypotheses.label}{label of the hypothesis}
}
\description{
given the adj.matrix, return the incoming and outgoing connections for any hypothesis
}

