% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.evaluate.cycles}
\alias{hypothesis.evaluate.cycles}
\title{hypothesis.evaluate.cycles}
\usage{
hypothesis.evaluate.cycles(data, adj.matrix, hypotheses.labels, weights.matrix)
}
\arguments{
\item{data}{input genotypes and its hypotheses}

\item{adj.matrix}{adjacency matrix of the reconstructed topology}

\item{hypotheses.labels}{label of all the existing hypotheses}

\item{weights.matrix}{weights of any edge in the topology}
}
\description{
evaluate cycles involving any hypothesis
}

