% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{hypothesis.expand.connections}
\alias{hypothesis.expand.connections}
\title{hypothesis.expand.connections}
\usage{
hypothesis.expand.connections(label, events, incoming, outgoing, hnames,
  matomic, weights.matrix)
}
\arguments{
\item{label}{name of the hypothesis}

\item{events}{events in the hypothesis}

\item{incoming}{incoming connections}

\item{outgoing}{outgoing connections}

\item{hnames}{Hypothesis names}

\item{matomic}{Map to atomic events}

\item{weights.matrix}{weights of any edge in the topology}
}
\description{
expand and enumerate all the connections incoming or outgoing an hypothesis
}

