% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{import.genotypes}
\alias{import.genotypes}
\title{import.genotypes}
\usage{
import.genotypes(geno, event.type = "variant", color = "Darkgreen")
}
\arguments{
\item{geno}{Either a dataframe or a filename}

\item{event.type}{Any 1 in "geno" will be interpreted as a an observed alteration labeled with type "event.type"}

\item{color}{This is the color used for visualization of events labeled as of "event.type"}
}
\value{
A TRONCO compliant dataset
}
\description{
Import a matrix of 0/1 alterations as a TRONCO compliant dataset. Input "geno" can be either a dataframe or
a file name. In any case the dataframe or the table stored in the file must have a column for each altered
gene and a rows for each sample. Colnames will be used to determine gene names, if data is loaded from
file the first column will be assigned as rownames.
}

