% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{import.GISTIC}
\alias{import.GISTIC}
\title{import.GISTIC}
\usage{
import.GISTIC(x)
}
\arguments{
\item{x}{Either a dataframe or a filename}
}
\value{
A TRONCO compliant representation of the input CNAs.
}
\description{
Transform GISTIC scores for CNAs in a TRONCO compliant object. Input can be either a matrix, with columns
for each altered gene and rows for each sample; in this case colnames/rownames mut be provided. If input
is a character an attempt to load a table from file is performed. In this case the input table format
should be constitent with TCGA data for focal CNA; there should hence be: one column for each sample,
one row for each gene, a column Hugo_Symbol with every gene name and a column Entrez_Gene_Id with every
 gene\'s Entrez ID. A valid GISTIC score should be any value of: "Homozygous Loss" (-2), "Heterozygous
 Loss" (-1), "Low-level Gain" (+1), "High-level Gain" (+2).
}
\examples{
data(gistic)
gistic = import.GISTIC(gistic)
gistic = annotate.description(gistic, 'Example GISTIC')
oncoprint(gistic)
}

