% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{import.model}
\alias{import.model}
\title{import.model}
\usage{
import.model(tronco_object, model, model.name = "imported_model")
}
\arguments{
\item{tronco_object}{A TRONCO compliant object}

\item{model}{Either a dataframe or a filename}

\item{model.name}{Name of the imported model}
}
\value{
A TRONCO compliant object
}
\description{
Add an adjacency matrix as a model to a TRONCO compliant object.
Input model can be either a dataframe or a file name.
}
