% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/external.R
\name{import.mutex.groups}
\alias{import.mutex.groups}
\title{import mutex group}
\usage{
import.mutex.groups(file, fdr = 0.2, display = TRUE)
}
\arguments{
\item{file}{Mutex results ("ranked-groups.txt" file)}

\item{fdr}{cutoff for fdr}

\item{display}{print summary table of extracted groups}
}
\description{
import mutex group
}

