% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{intersect.datasets}
\alias{intersect.datasets}
\title{intersect.datasets}
\usage{
intersect.datasets(x, y, intersect.genomes = TRUE)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{y}{A TRONCO compliant dataset.}

\item{intersect.genomes}{If False -> just samples}
}
\value{
A TRONCO complian dataset.
}
\description{
Intersect samples and events of two dataset
}
\examples{
data(test_dataset)
}

