% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctness.R
\name{is.events.list}
\alias{is.events.list}
\title{is.events.list}
\usage{
is.events.list(x, y)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{y}{A TRONCO event list}
}
\description{
Check if y is a valid event list for x
}

