% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{is.logic.node.down}
\alias{is.logic.node.down}
\title{is logical node down}
\usage{
is.logic.node.down(node)
}
\arguments{
\item{node}{A node identifier}
}
\value{
boolean
}
\description{
Check if logic node down
}

