% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{is.logic.node}
\alias{is.logic.node}
\title{is logical node}
\usage{
is.logic.node(node)
}
\arguments{
\item{node}{A node identifier}
}
\value{
boolean
}
\description{
Check if logic node down or up
}

