% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{is.logic.node.up}
\alias{is.logic.node.up}
\title{is logical node up}
\usage{
is.logic.node.up(node)
}
\arguments{
\item{node}{A node identifier}
}
\value{
boolean
}
\description{
Check if logic node up
}

