% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{join.events}
\alias{join.events}
\title{join.events}
\usage{
join.events(x, ..., new.event, new.type, event.color)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{...}{A list of events to merge}

\item{new.event}{The name of the resultant event}

\item{new.type}{The type of the new event}

\item{event.color}{The color of the new event}
}
\value{
A TRONCO compliant dataset.
}
\description{
Merge a list of events in an unique event
}
\examples{
data(muts)
dataset = join.events(muts, 'G1', 'G2', new.event='test', new.type='banana', event.color='yellow')

}
