% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{likertToClus}
\alias{likertToClus}
\title{Calculate the likert}
\usage{
likertToClus(cluster_result, sample_stage, cluster_prefix = "",
  sample_prefix = "")
}
\arguments{
\item{cluster_result}{Clustering result eg: [1, 2, 1, 3 ,3]}

\item{sample_stage}{Stage in which the sample is eg: [3, 3, 1, 2 ,2]}

\item{cluster_prefix}{Prefix to prefend to cluster data}

\item{sample_prefix}{Prefix to prefend to stage data}
}
\description{
Calculate the likert
}

