% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maf}
\alias{maf}
\title{MAF example data}
\format{Manual Annotated Format}
\source{
fake data
}
\usage{
data(maf)
}
\description{
This dataset contains a standard MAF input for TRONCO
}
\author{
Luca De Sano
}

