% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{merge.events}
\alias{merge.events}
\title{merge.events}
\usage{
\method{merge}{events}(x, events, new.event, new.type, event.color)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{events}{A list of events to merge}

\item{new.event}{The name of the resultant event}

\item{new.type}{The type of the new event}

\item{event.color}{The color of the new event}
}
\value{
A TRONCO compliant dataset.
}
\description{
Merge a list of events in an unique event
}
\examples{
data(muts)
dataset = merge.events(test_dataset, c('G1', 'G2'))
}

