% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muts}
\alias{muts}
\title{Simple mutation dataset}
\format{TRONCO compliant dataset}
\source{
fake data
}
\usage{
data(muts)
}
\description{
A simple mutation dataset without hypotheses
}
\author{
Luca De Sano
}
