% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{nameToKey}
\alias{nameToKey}
\title{nameToKey}
\usage{
nameToKey(x, name)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{name}{A intelligible event name}
}
\value{
A TRONCO dataset key name
}
\description{
Convert to key an intelligible event names, e.g., change 'Mutation KRAS' in G23.
If a name is not found, an error is raised!
}
\examples{
data(test_model)
adj_matrix = as.adj.matrix(test_model, events=as.events(test_model)[5:15,])$bic

}
