% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{nevents}
\alias{nevents}
\title{nevents}
\usage{
nevents(x, genes = NA, types = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{genes}{The genes to consider, if NA all available genes are used.}

\item{types}{The types of events to consider, if NA all available types are used.}
}
\value{
The number of events in the dataset involving a certain gene or type of event.
}
\description{
Return the number of events in the dataset involving a certain gene or type of event.
}
\examples{
data(test_dataset)
nevents(test_dataset)

}
