% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{ngenes}
\alias{ngenes}
\title{ngenes}
\usage{
ngenes(x, types = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{types}{The types of events to consider, if NA all available types are used.}
}
\value{
The number of genes in the dataset involving a certain type of event.
}
\description{
Return the number of genes in the dataset involving a certain type of event.
}
\examples{
data(test_dataset)
ngenes(test_dataset)
}

