% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{nhypotheses}
\alias{nhypotheses}
\title{Return the number of hypotheses in the dataset}
\usage{
nhypotheses(x)
}
\arguments{
\item{x}{the dataset.}
}
\description{
Return the number of hypotheses in the dataset
}
\examples{
data(test_dataset)
nhypotheses(test_dataset)
}

