% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{npatterns}
\alias{npatterns}
\title{Return the number of patterns in the dataset}
\usage{
npatterns(x)
}
\arguments{
\item{x}{the dataset.}
}
\description{
Return the number of patterns in the dataset
}
\examples{
data(test_dataset)
npatterns(test_dataset)


}
