% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{nsamples}
\alias{nsamples}
\title{nsamples}
\usage{
nsamples(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The number of samples in the dataset.
}
\description{
Return the number of samples in the dataset.
}
\examples{
data(test_dataset)
nsamples(test_dataset)
}

