% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{ntypes}
\alias{ntypes}
\title{Return the number of types in the dataset involving a certain set of genes}
\usage{
ntypes(x, genes = NA)
}
\arguments{
\item{x}{the dataset.}

\item{genes}{the genes to consider, if NA all available ones are used.}
}
\description{
Return the number of types in the dataset involving a certain set of genes
}

