% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{oncoprint}
\alias{oncoprint}
\title{oncoprint}
\usage{
oncoprint(x, excl.sort = TRUE, col.cluster = FALSE, row.cluster = FALSE,
  file = NA, ann.stage = has.stages(x), ann.hits = TRUE,
  stage.color = "YlOrRd", hits.color = "Purples",
  null.color = "lightgray", border.color = "white", font.size = 7,
  font.column = 3, font.row = NA, title = paste("Dataset (genotypes)"),
  sample.id = FALSE, hide.zeroes = FALSE, legend = TRUE, legend.cex = 1,
  cellwidth = NA, cellheight = NA, group.by.label = FALSE,
  group.samples = NA, pathways = NA, pathways.color = "Set1",
  txt.stats = paste("n = ", nsamples(x), "\\nm = ", nevents(x), "\\n|G| = ",
  ngenes(x), sep = ""), ...)
}
\arguments{
\item{col.cluster}{A number}

\item{hits.color}{= 'Purples' Color Palette to use with score}

\item{font.column}{= 3}

\item{sample.id}{= F Show sample name at the bottom of the heatmap}

\item{hide.zeroes}{= F Hide events without mutations}

\item{excl.soft}{A number}

\item{row.cluster=FALSE}{}

\item{device.new=FALSE}{}

\item{file=NA}{}

\item{ann.stage=TRUE}{Show information about stage classification}

\item{ann.hits=TRUE}{Show information about the hits in a gene}

\item{stage.color='YlOrRd'}{Color Palette to use with stage}

\item{null.color='darkgray'}{Background color}

\item{border.color='white'}{}

\item{font.size=7}{}

\item{title=}{paste('Genotypes')}
}
\description{
oncoprint
}

