% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{OR}
\alias{OR}
\title{OR}
\usage{
OR(...)
}
\arguments{
\item{...}{Atoms of the soft exclusive pattern given either as labels or as partielly lifted vectors.}
}
\value{
Vector to be added to the lifted genotype resolving the soft exclusive pattern
}
\description{
OR hypothesis
}

