% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{order.frequency}
\alias{order.frequency}
\title{order.frequency}
\usage{
order.frequency(x, decreasing = TRUE)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{decreasing}{Inverse order. Default TRUE}
}
\value{
A TRONCO compliant dataset with the internal genotypes sorted according to event frequency.
}
\description{
Sort the internal genotypes according to event frequency.
}
\examples{
data(test_dataset)
order.frequency(test_dataset)

}

