% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ov.cgh.R
\docType{data}
\name{ov.cgh}
\alias{ov.cgh}
\title{Ovarian cancer CGH data}
\format{A data frame with 87 observations on 7 variables.}
\source{
\url{http://www.ncbi.nlm.nih.gov/sky/}
}
\description{
This is a data set obtained using the comparative genomic
hybridization technique (CGH) on sam- ples from papillary serous
cystadenocarcinoma of the ovary. Only the seven most commonly
occurring events are given.
}
\details{
The CGH technique uses fluorescent staining to detect abnormal
(increased or decreased) number of DNA copies. Often the results
are reported as a gain or loss on a certain arm, without further
 distinction for specific regions. It is common to denote a change
 in DNA copy number on a specific chromosome arm by prefixing a "-"
 sign for decrease and a "+" for increase. Thus, say, -3q denotes
 abnormally low DNA copy number on the q arm of the 3rd chromosome.
}

