% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{pathway.visualization}
\alias{pathway.visualization}
\title{pathway.visualization}
\usage{
pathway.visualization(x, title = paste("Pathways:", paste(names(pathways),
  collapse = ", ", sep = "")), file, pathways.color = "Set2",
  aggregate.pathways, pathways, ...)
}
\arguments{
\item{x}{A TRONCO complian dataset}

\item{title}{Plot title}

\item{file}{To generate a PDF a filename have to be given}

\item{pathways.color}{A RColorBrewer color palette}
}
\value{
plot information
}
\description{
Visualise pathways informations
}

