% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patterns.R
\name{pattern.events}
\alias{pattern.events}
\title{Return all events involving certain genes and types}
\usage{
pattern.events(x, hypothesis)
}
\description{
Return all events involving certain genes and types
}

