% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{perform.likelihood.fit}
\alias{perform.likelihood.fit}
\title{perform.likelihood.fit}
\usage{
perform.likelihood.fit(dataset, adj.matrix, command, regularization)
}
\arguments{
\item{dataset}{a valid dataset}

\item{adj.matrix}{the adjacency matrix of the prima facie causes}

\item{command}{type of search, either hill climbing (hc) or tabu (tabu)}

\item{regularization}{regularization term to be used in the likelihood fit}
}
\value{
topology: the adjacency matrix of both the prima facie and causal topologies
}
\description{
reconstruct the best causal topology by likelihood fit
}

