% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection.R
\name{rank.recurrents}
\alias{rank.recurrents}
\title{rank.recurrents}
\usage{
rank.recurrents(x, n)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{n}{The number of events to rank}
}
\value{
the first n recurrent events
}
\description{
Return the first n recurrent events
}
\examples{
data(test_dataset)
dataset = rank.recurrents(test_dataset, 10)
}

