% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{remove.cycles}
\alias{remove.cycles}
\title{remove.cycles}
\usage{
remove.cycles(adj.matrix, weights.temporal.priority, weights.matrix,
  not.ordered, hypotheses = NA, silent)
}
\arguments{
\item{adj.matrix}{adjacency matrix of the topology}

\item{weights.temporal.priority}{weighted matrix to be used to remove the cycles involving atomic events}

\item{weights.matrix}{weighted matrix to be used to remove the cycles involving hypotheses}

\item{not.ordered}{list of the nodes to be orderd}

\item{hypotheses}{hypotheses to evaluate potential cycles}
}
\value{
acyclic.topology: structure representing the best acyclic topology
}
\description{
remove any cycle from a given cyclic topology
}

