% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{rename.gene}
\alias{rename.gene}
\title{rename.gene}
\usage{
rename.gene(x, old.name, new.name)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{old.name}{The name of the gene to rename.}

\item{new.name}{The new name}
}
\value{
A TRONCO complian dataset.
}
\description{
Rename a gene
}
\examples{
data(test_dataset)
test_dataset = rename.gene(test_dataset, 'TET2', 'gene x')
}

