% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{rename.type}
\alias{rename.type}
\title{rename.type}
\usage{
rename.type(x, old.name, new.name)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{old.name}{The type of event to rename.}

\item{new.name}{The new name}
}
\value{
A TRONCO complian dataset.
}
\description{
Rename an event type
}
\examples{
data(test_dataset)
test_dataset = rename.type(test_dataset, 'ins_del', 'deletion')

}

