% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection.R
\name{samples.selection}
\alias{samples.selection}
\title{samples.selection}
\usage{
samples.selection(x, samples)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{samples}{A list of samples}
}
\value{
A TRONCO compliant dataset.
}
\description{
Filter a dataset based on selected samples id
}
\examples{
data(test_dataset)
dataset = samples.selection(test_dataset, c('patient 1', 'patient 2'))
}

