% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{sbind}
\alias{sbind}
\title{sbind}
\usage{
sbind(...)
}
\arguments{
\item{...}{the input datasets}
}
\value{
A TRONCO complian dataset.
}
\description{
Binds samples from one or more datasets, which must be defined over the same set of events
}
