% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{show}
\alias{show}
\title{show}
\usage{
show(x, view = 10)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{view}{The firse \code{view} events are shown via \code{head}.}
}
\description{
Print to console a short report of a dataset 'x', which should be
a TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
show(test_dataset)

}

