% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{sort.by.frequency}
\alias{sort.by.frequency}
\title{sort.by.frequency}
\usage{
\method{sort}{by.frequency}(x, decreasing = TRUE, ...)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{decreasing}{Inverse order. Default TRUE}

\item{...}{just for compatibility}
}
\value{
A TRONCO compliant dataset with the internal genotypes sorted according to event frequency.
}
\description{
Sort the internal genotypes according to event frequency.
}
\examples{
data(test_dataset)
sort.by.frequency(test_dataset)

}

