% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{ssplit}
\alias{ssplit}
\title{ssplit}
\usage{
ssplit(x, clusters, idx = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{clusters}{A list of clusters. Rownames must match samples list of x}

\item{idx}{ID of a specific group present in stages. If NA all groups will be extracted}
}
\value{
A TRONCO compliant dataset.
}
\description{
Split cohort (samples) into groups, return either all groups or a specific group.
}
