% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stat.eloss}
\alias{stat.eloss}
\title{stat.eloss}
\usage{
stat.eloss(data, regularization = "bic", runs = 10, k = 10)
}
\arguments{
\item{data}{A reconstructed model (the output of tronco.capri or tronco.caprese)}

\item{regularization}{The name of the selected regularization (default: "bic")}

\item{runs}{a positive integer number, the number of times cross-validation will be run}

\item{k}{a positive integer number, the number of groups into which the data will be split}
}
\description{
Perform a k-fold cross-validation using the function bn.cv.
}
\examples{
data(test_model)
stat.eloss(test_model)

}

