% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stat.prederr}
\alias{stat.prederr}
\title{stat.prederr}
\usage{
stat.prederr(data, regularization = "bic", nodes = as.events(data,
  keysToNames = TRUE), runs = 10)
}
\arguments{
\item{data}{A reconstructed model (the output of tronco.capri or tronco.caprese)}

\item{regularization}{The name of the selected regularization (default: "bic")}

\item{nodes}{a list of event}

\item{runs}{a positive integer number, the number of times cross-validation will be run}
}
\description{
Perform a k-fold cross-validation (with k = 10) using the function bn.cv
and scan every node to estimate its prediction error.
}
\examples{
data(test_model)
stat.prederr(test_model)

}

