% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/external.R
\name{TCGA.map.clinical.data}
\alias{TCGA.map.clinical.data}
\title{TCGA.map.clinical.data}
\usage{
TCGA.map.clinical.data(file, sep = "\\t", column.samples, column.map)
}
\arguments{
\item{file}{TODO}

\item{sep}{file delimiter}

\item{column.samples}{TODO}

\item{column.map}{TODO}
}
\value{
a map
}
\description{
TODO
}

