% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/external.R
\name{TCGA.multiple.samples}
\alias{TCGA.multiple.samples}
\title{TCGA.multiple.samples}
\usage{
TCGA.multiple.samples(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A list of barcodes. NA if no duplicated barcode is found
}
\description{
Check if there are multiple sample in x, according to TCGA barcodes naming
}

