% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{TCGA.shorten.barcodes}
\alias{TCGA.shorten.barcodes}
\title{TCGA.shorten.barcodes}
\usage{
TCGA.shorten.barcodes(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A TRONCO compliant dataset
}
\description{
Keep only the first 12 character of samples barcode if there are no duplicates
}

