% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dataset}
\alias{test_dataset}
\title{A complete dataset with hypotheses}
\format{TRONCO compliant dataset}
\source{
fake data
}
\usage{
data(test_dataset)
}
\value{
A standard TRONCO object
}
\description{
This dataset contains a complete test dataset
}
\author{
Luca De Sano
}
