% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dataset_no_hypos}
\alias{test_dataset_no_hypos}
\title{A complete dataset}
\format{
TRONCO compliant dataset
}
\source{
fake data
}
\usage{
data(test_dataset_no_hypos)
}
\value{
A standard TRONCO object
}
\description{
This dataset contains a complete test dataset
}
\author{
Luca De Sano
}
