% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_model}
\alias{test_model}
\title{A complete dataset with a reconstructed model}
\format{TRONCO compliant dataset}
\source{
fake data
}
\usage{
data(test_dataset)
}
\description{
This dataset contains a model reconstructed with CAPRI
}
\author{
Luca De Sano
}

