% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.caprese}
\alias{tronco.caprese}
\title{tronco caprese}
\usage{
tronco.caprese(data, lambda = 0.5, do.estimation = FALSE, silent = FALSE)
}
\arguments{
\item{data}{A TRONCO compliant dataset.}

\item{lambda}{TODO}

\item{do.estimation}{TODO}

\item{silent}{TODO}
}
\value{
A TRONCO compliant object with reconstructed model
}
\description{
Reconstruc a progression model using CAPRESE algorithm
}
\examples{
data(maf)
mutations = import.MAF(maf)
recon = tronco.caprese(mutations)
tronco.plot(recon)
}

