% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.capri}
\alias{tronco.capri}
\title{tronco capri}
\usage{
tronco.capri(data, command = "hc", regularization = c("bic", "aic"),
  do.boot = TRUE, nboot = 100, pvalue = 0.05, min.boot = 3,
  min.stat = TRUE, boot.seed = NULL, do.estimation = FALSE,
  silent = FALSE)
}
\arguments{
\item{data}{A TRONCO compliant dataset.}

\item{do.estimation}{TODO}

\item{silent}{TODO}

\item{lambda}{TODO}
}
\value{
A TRONCO compliant object with reconstructed model
}
\description{
Reconstruc a progression model using CAPRI algorithm
}
\examples{
data(maf)
mutations = import.MAF(maf)
recon = tronco.capri(mutations)
tronco.plot(recon)
}

