% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.estimation}
\alias{tronco.estimation}
\title{tronco.estimation}
\usage{
tronco.estimation(reconstruction, error.rates = NA)
}
\arguments{
\item{reconstruction}{A TRONCO compliant dataset with a reconstructed model associated.}

\item{error.rates}{todo}
}
\value{
A TRONCO compliant object with reconstructed model and estimations
}
\description{
todo
}
\examples{
data(test_model)
recon = tronco.estimation(test_model)
}

