% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{tronco.kfold.eloss}
\alias{tronco.kfold.eloss}
\title{tronco.kfold.eloss}
\usage{
tronco.kfold.eloss(x, regularization = "bic", runs = 10, k = 10)
}
\arguments{
\item{regularization}{The name of the selected regularization (default: "bic")}

\item{runs}{a positive integer number, the number of times cross-validation will be run}

\item{k}{a positive integer number, the number of groups into which the data will be split}

\item{data}{A reconstructed model (the output of tronco.capri or tronco.caprese)}
}
\description{
Perform a k-fold cross-validation using the function bn.cv.
}
\examples{
data(test_model)
stat.eloss(test_model)

}

