% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{tronco.kfold.prederr}
\alias{tronco.kfold.prederr}
\title{tronco.kfold.prederr}
\usage{
tronco.kfold.prederr(x, regularization = as.parameters(x)$regularization,
  events = as.events(x), runs = 10, k = 10, cores.ratio = 1,
  verbose = FALSE)
}
\arguments{
\item{x}{A reconstructed model (the output of tronco.capri)}

\item{regularization}{The name of the selected regularization (default: "bic")}

\item{events}{a list of event}

\item{runs}{a positive integer number, the number of times cross-validation will be run}

\item{k}{a positive integer number, the number of groups into which the data will be split}

\item{cores.ratio}{Percentage of cores to use. coresRate * (numCores - 1)}
}
\description{
Perform a k-fold cross-validation using the function bn.cv
and scan every node to estimate its prediction error.
}
\examples{
data(test_model)
tronco.kfold.prederr(test_model)

}

