% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{tronco.pattern.plot}
\alias{tronco.pattern.plot}
\title{tronco.pattern.plot}
\usage{
tronco.pattern.plot(
  x,
  group = as.events(x),
  to,
  gap.cex = 1,
  legend.cex = 1,
  label.cex = 1,
  title = paste(to[1], to[2]),
  mode = "barplot"
)
}
\arguments{
\item{x}{A TRONCO compliant dataset}

\item{group}{A list of events (see as.events() for details)}

\item{to}{A target event}

\item{gap.cex}{cex parameter for gap}

\item{legend.cex}{cex parameter for legend}

\item{label.cex}{cex parameter for label}

\item{title}{title}

\item{mode}{can be 'circos' or 'barplot'}
}
\description{
tronco.pattern.plot : plot a genotype
}
