% Generated by roxygen2 (4.0.2): do not edit by hand
\name{types.add}
\alias{types.add}
\title{add a new type of event (e.g., missense point mutation)}
\usage{
types.add(type.name, color.name)
}
\arguments{
\item{type.name}{The type label. All type labels are strings.}

\item{color.name}{The type color. All R's color definitions are allowed.}
}
\description{
\code{types.add} sets a global data frame 'types' that contains all types defined.  Types can be added and refined incrementally, in any order.
}
\details{
\code{types.add} defines a type of event considered  at a time. If the type was previously defined, its definition is updated to keep track of its last definition. A consistency check is performed to ensure that the type is valid. Types must be defined before events are loaded.
}
\examples{
types.add("gain", "red")
}

