% Generated by roxygen2 (4.0.2): do not edit by hand
\name{types.load}
\alias{types.load}
\title{load a set of types from file}
\usage{
types.load(data.input)
}
\arguments{
\item{data.input}{The input file path or a dataset to be validated.}
}
\description{
\code{types.load} sets a global data frame 'types' that contains all type definitions found in a specified file or dataset
to be validated.
}
\details{
\code{types.load} allows to load type definitions from a given file path. The file which contains
all the definitions must be structured as a csv file. All definitions are couple of values
type name and color name as shown below:

typeName, colorName
...     , ...
}
\seealso{
\code{\link{types.add}}
}

