% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{types}
\alias{types}
\title{Types collection for Ovarian cancer CGH data}
\format{An example with 2 types}
\description{
This example contains a collection of types associeted to the
Ovarian cancer CGH dataset
}

