% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.algorithm.R
\name{verify.parents}
\alias{verify.parents}
\title{verify.parents}
\usage{
verify.parents(best.parents, marginal.probs, joint.probs)
}
\arguments{
\item{best.parents}{best edges to be verified}

\item{marginal.probs}{observed marginal probabilities}

\item{joint.probs}{observed joint probabilities}
}
\value{
best.parents: list of the best valid parents
}
\description{
verify the independent progression filter
}

