% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{verify.probability.raising.do.boot}
\alias{verify.probability.raising.do.boot}
\title{verify.probability.raising.do.boot}
\usage{
verify.probability.raising.do.boot(prima.facie.model.distributions,
  prima.facie.null.distributions, pvalue, adj.matrix, edge.confidence.matrix)
}
\arguments{
\item{prima.facie.model.distributions}{distributions of the prima facie model}

\item{prima.facie.null.distributions}{distributions of the prima facie null}

\item{pvalue}{minimum pvalue for the Mann-Whitney U tests to be significant}

\item{adj.matrix}{adjacency matrix of the topology}

\item{edge.confidence.matrix}{matrix of the confidence of each edge}
}
\value{
probability.raising: list describing the causes where probability raising is verified
}
\description{
verify the probability raising condition
}

