% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{verify.probability.raising.no.boot}
\alias{verify.probability.raising.no.boot}
\title{verify.probability.raising.no.boot}
\usage{
verify.probability.raising.no.boot(prima.facie.model, prima.facie.null,
  adj.matrix, edge.confidence.matrix)
}
\arguments{
\item{prima.facie.model}{prima facie model}

\item{prima.facie.null}{prima facie null}

\item{adj.matrix}{adjacency matrix of the topology}

\item{edge.confidence.matrix}{matrix of the confidence of each edge}
}
\value{
probability.raising: adjacency matrix where temporal priority is verified
}
\description{
verify the probability raising condition without bootstrap
}

