% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{verify.temporal.priority.do.boot}
\alias{verify.temporal.priority.do.boot}
\title{verify.temporal.priority.do.boot}
\usage{
verify.temporal.priority.do.boot(marginal.probs.distributions, pvalue,
  adj.matrix, edge.confidence.matrix)
}
\arguments{
\item{marginal.probs.distributions}{distributions of the bootstrapped marginal probabilities}

\item{pvalue}{minimum pvalue for the Mann-Whitney U tests to be significant}

\item{adj.matrix}{adjacency matrix of the topology}

\item{edge.confidence.matrix}{matrix of the confidence of each edge}
}
\value{
temporal.priority: list describing the causes where temporal priority is verified
}
\description{
verify the temporal priority condition with bootstrap
}

