% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{verify.temporal.priority.no.boot}
\alias{verify.temporal.priority.no.boot}
\title{verify.temporal.priority.no.boot}
\usage{
verify.temporal.priority.no.boot(marginal.probs, adj.matrix,
  edge.confidence.matrix)
}
\arguments{
\item{marginal.probs}{marginal probabilities}

\item{adj.matrix}{adjacency matrix of the topology}

\item{edge.confidence.matrix}{matrix of the confidence of each edge}
}
\value{
temporal.priority: adjacency matrix where temporal priority is verified
}
\description{
verify the temporal priority condition without bootstrap
}

