% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/which.functions.R
\name{which.samples}
\alias{which.samples}
\title{which.samples}
\usage{
which.samples(x, gene, type, neg = FALSE)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{gene}{A list of gene names}

\item{type}{A list of types}

\item{neg}{If FALSE return the list, if TRUE return as.samples() - list}
}
\value{
A list of sample
}
\description{
Return a list of samples with specified alteration
}
\examples{
data(test_dataset)
which.samples(test_dataset, 'TET2', 'ins_del')
which.samples(test_dataset, 'TET2', 'ins_del', neg=TRUE)
}

