% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.hypotheses.R
\name{XOR}
\alias{XOR}
\title{XOR}
\usage{
XOR(...)
}
\arguments{
\item{...}{Atoms of the hard exclusive pattern given either as labels or as partielly lifted vectors.}
}
\value{
Vector to be added to the lifted genotype resolving the hard exclusive pattern
}
\description{
XOR hypothesis
}

