% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{as.bnlearn.network}
\alias{as.bnlearn.network}
\title{as.bnlearn.network}
\usage{
as.bnlearn.network(obj, regularization = "bic", makeValid = TRUE)
}
\arguments{
\item{regularization}{The name of the selected regularization (default: "bic")}

\item{makeValid}{Transform the bootstrapped data into a valid 2-categories input data}

\item{data}{A reconstructed model (the output of tronco.capri or tronco.caprese)}
}
\description{
Convert a TRONCO object in a Bnlearn network.
}
\examples{
data(test_model)
as.bnlearn.network(test_model)

}

