% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{as.bnlearn.network}
\alias{as.bnlearn.network}
\title{as.bnlearn.network}
\usage{
as.bnlearn.network(data, regularization = "bic")
}
\arguments{
\item{data}{A reconstructed model (the output of tronco.capri or tronco.caprese)}

\item{regularization}{The name of the selected regularization (default: "bic")}
}
\description{
Convert a TRONCO object in a Bnlean network
}
\examples{
data(test_model)
as.bnlearn.network(test_model)

}

