% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.events.hypotheses}
\alias{as.events.hypotheses}
\title{Return all events involving certain genes and types.}
\usage{
as.events.hypotheses(x, hypotheses = NULL)
}
\arguments{
\item{x:}{the dataset.}

\item{hypotheses:}{}
}
\description{
Return all events involving certain genes and types.
}

