% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.genotypes}
\alias{as.genotypes}
\title{as genotypes}
\usage{
as.genotypes(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A TRONCO genotypes matrix.
}
\description{
Return all genotypes for input 'x', which should be a TRONCO compliant dataset
see \code{is.compliant}.
Function \code{keysToNames} can be used to translate colnames to events.
}
\examples{
data(maf)
mutations = import.MAF(maf, sep=';')
keysToNames(mutations, as.genotypes(mutations))
}

