% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as.genotypes.test}
\alias{as.genotypes.test}
\title{as genotypes matrix}
\format{matrix}
\source{
da mettere
}
\usage{
data(as.genotypes.test)
}
\description{
This data set list ...
}
\author{
Luca De Sano
}

