% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.hypotheses}
\alias{as.hypotheses}
\title{as.hypotheses}
\usage{
as.hypotheses(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The patterns in the dataset which constitute CAPRI's hypotheses.
}
\description{
Return the patterns in the dataset which constitute CAPRI's hypotheses.
RIDENOMINARLA CON PATTERNS?
}

