% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.hypotheses}
\alias{as.hypotheses}
\title{as.hypotheses}
\usage{
as.hypotheses(x, cause = NA, effect = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{cause}{A list of genes to use as causes}

\item{effect}{A list of genes to use as effects}
}
\value{
The hypotheses in the dataset which constitute CAPRI's hypotheses.
}
\description{
Return the hypotheses in the dataset which constitute CAPRI's hypotheses.
}
\examples{
data(test_dataset)
as.hypotheses(test_dataset)

}

