% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.hypotheses}
\alias{as.hypotheses}
\title{as.hypotheses}
\usage{
as.hypotheses(x, cause = NULL, effect = NULL)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The hypotheses in the dataset which constitute CAPRI's hypotheses.
}
\description{
Return the hypotheses in the dataset which constitute CAPRI's hypotheses.
}
\examples{
data(test_dataset)
as.hypotheses(test_dataset)
}

