% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R, R/patterns.R
\name{as.patterns}
\alias{as.patterns}
\title{as.patterns}
\usage{
as.patterns(x)

as.patterns(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The patterns in the dataset which constitute CAPRI's hypotheses.
}
\description{
Return the patterns in the dataset which constitute CAPRI's hypotheses.

Return the names of the patterns in the dataset
}

