% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{as.samples}
\alias{as.samples}
\title{as.samples}
\usage{
as.samples(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A vector of sample IDs
}
\description{
Return all sample IDs for input 'x', which should be a TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
as.samples(test_dataset)

}

