% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caprese.bootstrap.R
\name{bootstrap.caprese}
\alias{bootstrap.caprese}
\title{bootstrap.caprese}
\usage{
bootstrap.caprese(dataset, lambda, do.estimation, silent, reconstruction,
  command = "non-parametric", nboot = 100, bootstrap.statistics = list())
}
\arguments{
\item{dataset}{a dataset describing a progressive phenomenon}

\item{lambda}{shrinkage parameter (value in [0,1])}

\item{do.estimation}{todo}

\item{silent}{todo}

\item{reconstruction}{todo}

\item{command}{todo}

\item{nboot}{number of bootstrap resampling to be performed}

\item{bootstrap.statistics}{todo}
}
\value{
bootstrap.statistics: statistics of the bootstrap
}
\description{
perform non-parametric or parametric bootstrap to evalutate the confidence of the reconstruction
}

