% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.hypothesis}
\alias{delete.hypothesis}
\title{Delete of a single hypothesis}
\usage{
delete.hypothesis(x, event = NULL, cause = NULL, effect = NULL)
}
\description{
Delete of a single hypothesis
}

