% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{delete.hypothesis}
\alias{delete.hypothesis}
\title{delete.hypothesis}
\usage{
\method{delete}{hypothesis}(x, event = NA, cause = NA, effect = NA)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{event}{Can be an event or pattern name}

\item{cause}{Can be an event or pattern name}

\item{effect}{Can be an event or pattern name}
}
\value{
A TRONCO complian dataset.
}
\description{
Delete an hypothesis from the dataset based on a selected event.
Check if the selected event exist in the dataset and delete his associated hypothesis
}
\examples{
data(test_dataset)
delete.hypothesis(test_dataset, event='TET2')
delete.hypothesis(test_dataset, cause='EZH2')
delete.hypothesis(test_dataset, event='XOR_EZH2')

}

