% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{export.nbs.input}
\alias{export.nbs.input}
\title{export.nbs.input}
\usage{
export.nbs.input(x, map_hugo_entrez, file = "tronco_to_nbs.mat")
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{map_hugo_entrez}{Hugo_Symbol-Entrez_Gene_Id map}

\item{file}{output file name}
}
\description{
Create a .mat file which can be used with NBS clustering
(ref: http://chianti.ucsd.edu/~mhofree/wordpress/?page_id=26)
}
