% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capri.algorithm.R
\name{get.prima.facie.parents.no.boot}
\alias{get.prima.facie.parents.no.boot}
\title{get.prima.facie.parents.no.boot}
\usage{
get.prima.facie.parents.no.boot(dataset, hypotheses, adj.matrix, silent)
}
\arguments{
\item{dataset}{a valid dataset}

\item{hypotheses}{hypotheses object associated to dataset}

\item{adj.matrix}{adjacency matrix of the initially valid edges}

\item{silent}{Should I be verbose?}
}
\value{
prima.facie.parents: list of the set (if any) of prima facie parents for each node
}
\description{
select the set of the prima facie parents (without bootstrap) for each node based on Suppes' definition of causation
}

