% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{has.model}
\alias{has.model}
\title{has.model}
\usage{
has.model(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
TRUE if there is a reconstructed model in \code{x}.
}
\description{
Return true if there is a reconstructed model in the TRONCO dataset 'x', which should be
a TRONCO compliant dataset - see \code{is.compliant}.
}
\examples{
data(test_dataset)
has.model(test_dataset)

}
