% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{import.MAF}
\alias{import.MAF}
\title{import.MAF}
\usage{
import.MAF(file, sep = "\\t", is.TCGA = TRUE)
}
\arguments{
\item{file}{MAF filename}

\item{sep}{MAF separator, default "\t"}

\item{is.TCGA}{TRUE if this MAF is from TCGA; thus its sample codenames can be interpreted}
}
\value{
A TRONCO compliant representation of the input MAF
}
\description{
Import mutation profiles from a Manual Annotation Format (MAF) file. All mutations are aggregated as a
unique event type labeled "Mutation" and assigned a color according to the default of function
\code{import.genotypes}. If this is a TCGA MAF file check for multiple samples per patient is performed
and a warning is raised if these occurr.
}
\examples{
data(maf)
mutations = import.MAF(maf)
mutations = annotate.name(mutations, 'Example MAF')
mutations = TCGA.shorten.barcodes(mutations)
oncoprint(mutations)
}

