% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{import.mutex.groups}
\alias{import.mutex.groups}
\title{import.mutex.groups}
\usage{
import.mutex.groups(file, fdr = 0.2, display = TRUE)
}
\arguments{
\item{file}{Mutex results ("ranked-groups.txt" file)}

\item{fdr}{cutoff for fdr}

\item{display}{print summary table of extracted groups}
}
\description{
Create a list of unique Mutex groups for a given fdr cutoff
current Mutex version is Jan 8, 2015
(ref: https://code.google.com/p/mutex/ )
}

