% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{merge.types}
\alias{merge.types}
\title{merge.types}
\usage{
\method{merge}{types}(x, ..., new.type = "new.type", new.color = "khaki")
}
\arguments{
\item{x}{A TRONCO compliant dataset.}

\item{...}{type to merge}

\item{new.type}{label for the new type to create}

\item{new.color}{color for the new type to create}
}
\value{
A TRONCO compliant dataset.
}
\description{
For an input dataset merge all the events of two or more distincit types
(e.g., say that missense and indel mutations are events
of a unique "mutation" type)
}
\examples{
data(test_dataset)
dataset = merge.types(test_dataset, 'ins_del', 'missense_point_mutations')
dataset = merge.types(test_dataset, 'ins_del', 'missense_point_mutations', 'mut', 'green')
}

