% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R, R/patterns.R
\name{npatterns}
\alias{npatterns}
\title{Return the number of patterns in the dataset}
\usage{
npatterns(x)

npatterns(x)
}
\arguments{
\item{x}{the dataset.}
}
\description{
Return the number of patterns in the dataset

Return the number of patterns in the dataset
}

