% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.functions.R
\name{ntypes}
\alias{ntypes}
\title{ntypes}
\usage{
ntypes(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
The number of types in the dataset.
}
\description{
Return the number of types in the dataset.
}
\examples{
data(test_dataset)
ntypes(test_dataset)

}

