% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\alias{reset.events}
\alias{reset.types}
\title{reset}
\usage{
reset.events()

reset.types()

reset()
}
\description{
A set of functions to reset events, types and data.values variables
}
\details{
\code{reset.events} Resets the events variable

\code{reset.types()} Resets the types variable

\code{reset()} Resets types, events and data.values variables
}
\examples{
reset.events()
reset.types()
reset()
}

