% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.functions.R
\name{sort.by.frequency}
\alias{sort.by.frequency}
\title{sort.by.frequency}
\usage{
\method{sort}{by.frequency}(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
Sort the internal genotypes according to event frequency.
}
\description{
Sort the internal genotypes according to event frequency.
}

