% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stage}
\alias{stage}
\title{Stage information for test_dataset}
\format{Vector of stages}
\source{
fake data
}
\usage{
data(stage)
}
\description{
This dataset contains stage information for patient in test_dataset
}
\author{
Luca De Sano
}

