% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.functions.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(x)
}
\arguments{
\item{x}{A TRONCO compliant dataset.}
}
\value{
A TRONCO compliant dataset.
}
\description{
Deletes all events which have frequency 0 in the dataset.
}
\examples{
data(test_dataset)
test_dataset = trim(test_dataset)

}
