% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.bootstrap}
\alias{tronco.bootstrap}
\title{tronco bootstrap}
\usage{
tronco.bootstrap(reconstruction, type = "non-parametric", nboot = 100,
  verbose = FALSE)
}
\arguments{
\item{reconstruction}{The output of tronco.capri or tronco.caprese}

\item{type}{TODO}

\item{nboot}{TODO}
}
\value{
A TRONCO compliant object with reconstructed model
}
\description{
Bootstrap a reconstructed progression model
}
\examples{
data(maf)
mutations = import.MAF(maf)
recon = tronco.capri(mutations)
boot = tronco.bootstrap(recon)
tronco.plot(boot)
}

