% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.bootstrap.R
\name{tronco.bootstrap}
\alias{tronco.bootstrap}
\title{perform bootstrap algorithm}
\usage{
tronco.bootstrap(topology, lambda = 0.5, type = "non-parametric",
  nboot = 1000)
}
\arguments{
\item{topology}{A topology returned by a reconstruction algorithm}

\item{lambda}{A lambda value, default is 0.5}

\item{type}{The type of bootstrap performed, parametric and non parametric types are available.
To specify wich type of bootstrap run type must be "parametric" or "non-parametric".}

\item{nboot}{Samplig value. The grater will be the nboot value the logehr time the
entire process will take to complete the computing}
}
\value{
A topology object with bootstrap informations added
}
\description{
\code{tronco.bootstrap} perform parametric and non-parametric bootstrap algorithms
}

