% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.bootstrap}
\alias{tronco.bootstrap}
\title{tronco bootstrap}
\usage{
tronco.bootstrap(reconstruction, type = "non-parametric", nboot = 100,
  verbose = FALSE)
}
\arguments{
\item{reconstruction}{The output of tronco.capri or tronco.caprese}

\item{type}{Parameter to define the type of sampling to be performed, e.g., non-parametric for uniform sampling.}

\item{nboot}{Number of bootstrap sampling to be performed when estimating the model confidence.}

\item{verbose}{Should I be verbose?}
}
\value{
A TRONCO compliant object with reconstructed model
}
\description{
Bootstrap a reconstructed progression model
}
\examples{
data(test_dataset)
recon = tronco.capri(test_dataset)
boot = tronco.bootstrap(recon, nboot=5)
tronco.plot(boot)
}

