% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tronco.bootstrap.show}
\alias{tronco.bootstrap.show}
\title{show bootstrapping results}
\usage{
tronco.bootstrap.show(topology)
}
\arguments{
\item{topology}{A topology returned by a reconstruction algorithm}
}
\description{
\code{tronco.bootstrap.show} show bootstrapping results. Requires that you already executed tronco.bootstrap
}

