% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.caprese}
\alias{tronco.caprese}
\title{tronco caprese}
\usage{
tronco.caprese(data, lambda = 0.5, do.estimation = FALSE, silent = FALSE)
}
\arguments{
\item{data}{A TRONCO compliant dataset.}

\item{lambda}{Coefficient to combine the raw estimate with a correction factor into a shrinkage estimator.}

\item{do.estimation}{A parameter to disable/enable the estimation of the error rates give the reconstructed model.}

\item{silent}{A parameter to disable/enable verbose messages.}
}
\value{
A TRONCO compliant object with reconstructed model
}
\description{
Reconstruct a progression model using CAPRESE algorithm
}
\examples{
data(test_dataset)
recon = tronco.caprese(test_dataset)
tronco.plot(recon)

}

