% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tronco.caprese}
\alias{tronco.caprese}
\title{runs CAPRESE algorithm}
\usage{
tronco.caprese(dataset, lambda = 0.5, verbose = FALSE)
}
\arguments{
\item{dataset}{The input dataset. Type: dataframe. The dataset given as input is the data.values data frame loaded by the \code{data} function.}

\item{lambda}{the real positive value of the shrinkage coefficient, required to range in [0, 1]. Its default value is 0.5, if unspecified.}

\item{verbose}{execute CAPRESE algorithm with verbose output to screen. Type: boolean, dafault: FALSE.}
}
\value{
an object containing the reconstructed topology and confidence values.
}
\description{
\code{tronco.caprese} executes the CAPRESE algorithm on the dataset \code{data.values} specified.
}
\details{
\code{tronco.caprese} executes the reconstruction of the topology, and computesg all the confidence measures defined in \code{confidence}.
}
\seealso{
\code{\link{data}}
}

