% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tronco.R
\name{tronco.plot}
\alias{tronco.plot}
\title{plot a progression model}
\usage{
tronco.plot(x, regularization = names(x$model), fontsize = NA, height = 2,
  width = 3, height.logic = 1, pf = FALSE, disconnected = FALSE,
  scale.nodes = NA, name = deparse(substitute(capri)), title = as.name(x),
  confidence = NA, p.min = x$parameters$pvalue, legend = TRUE,
  legend.cex = 1, edge.cex = 1, label.edge.size = NA, expand = TRUE,
  genes = NULL, relations.filter = NA, edge.color = "black",
  pathways.color = "Set1", file = NA, legend.pos = "bottom",
  pathways = NULL, lwd = 3, ...)
}
\description{
\code{tronco.plot} plots a progression model from a recostructed \code{curr.reconstruction}.
}

